feat: new lens selection menu
Took 5 hours 31 minutes

+++ src/main/java/ru/feytox/etherology/util/misc/RenderUtils.java
@@ -9,11 +9,11 @@ import net.minecraft.client.util.math.MatrixStack;
 @UtilityClass
 public class RenderUtils {
 
-    public static void renderTexture(MatrixStack matrices, float x0, float y0, int u, int v, int width, int height, int textureWidth, int textureHeight) {
+    public static void renderTexture(MatrixStack matrices, float x0, float y0, int u, int v, float width, float height, int textureWidth, int textureHeight) {
         renderTexture(matrices, x0, y0, u, v, width, height, textureWidth, textureHeight, textureWidth, textureHeight);
     }
 
-    public static void renderTexture(MatrixStack matrices, float x0, float y0, int u, int v, int width, int height, int regionWidth, int regionHeight, int textureWidth, int textureHeight) {
+    public static void renderTexture(MatrixStack matrices, float x0, float y0, int u, int v, float width, float height, int regionWidth, int regionHeight, int textureWidth, int textureHeight) {
         val matrix = matrices.peek().getPositionMatrix();
         float x1 = x0 + width;
         float y1 = y0 + height;
